requirements.txt
absl-py==1.4.0
annotated-types==0.5.0
anyio==3.7.1
apparmor==3.0.4
apturl==0.5.2
astunparse==1.6.3
Babel==2.8.0
bcc==0.18.0
bcrypt==3.2.0
blinker==1.4
Brlapi==0.8.3
cachetools==5.3.1
certifi==2020.6.20
chardet==4.0.0
click==8.0.3
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
distlib==0.3.6
distro==1.7.0
distro-info==1.1+ubuntu0.1
dnspython==2.1.0
duplicity==0.8.21
ecdsa==0.18.0
email-validator==2.0.0.post2
exceptiongroup==1.1.3
fastapi==0.103.0
fasteners==0.14.1
filelock==3.9.0
flatbuffers==23.5.26
fsspec==2023.10.0
future==0.18.2
gast==0.4.0
google-auth==2.23.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
gpg==1.16.0
greenlet==2.0.2
grpcio==1.58.0
h11==0.14.0
h5py==3.9.0
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
jeepney==0.7.1
Jinja2==3.0.3
joblib==1.3.2
keras==2.13.1
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
LibAppArmor==3.0.4
libclang==16.0.6
libcomps==0.1.15
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.4.4
MarkupSafe==2.1.3
monotonic==1.6
more-itertools==8.10.0
mpmath==1.3.0
netaddr==0.8.0
netifaces==0.11.0
networkx==3.2.1
norminette==3.3.54
numpy==1.24.3
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.0
olefile==0.46
opt-einsum==3.3.0
packaging==23.1
panda==0.3.1
pandas==2.1.3
paramiko==2.9.3
passlib==1.7.4
pexpect==4.8.0
Pillow==9.0.1
platformdirs==3.1.1
protobuf==4.24.3
proton-client==0.7.1
protonvpn-cli==3.13.0
protonvpn-gui==1.12.0
protonvpn-nm-lib==3.16.0
psutil==5.9.0
psycopg2-binary==2.9.7
ptyprocess==0.7.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycairo==1.20.1
pycups==2.0.1
pydantic==1.10.10
pydantic_core==2.0.2
pygame==2.5.2
Pygments==2.11.2
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyRFC3339==1.1
pyscopg2==66.0.2
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
python-gnupg==0.4.8
python-jose==3.3.0
python-multipart==0.0.6
pythondialog==3.5.1
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.25.1
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
rpm==4.17.0
rsa==4.9
scapy==2.5.0
scikit-learn==1.3.2
scipy==1.11.4
SecretStorage==3.3.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.20
starlette==0.27.0
sympy==1.12
systemd-python==234
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.2.0
threadpoolctl==3.2.0
torch==2.1.1
triton==2.1.0
typing_extensions==4.8.0
tzdata==2023.3
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
uvicorn==0.23.2
virtualenv==20.21.0
wadllib==1.3.6
Werkzeug==2.3.7
wrapt==1.15.0
xdg==5
xkit==0.0.0
zipp==1.0.0
